@@ -1,6 +1,9 @@
 using System;
+using System.Text;
 
-public readonly record struct CommandRouter( string Name, string Description, Func<string[], int> Handler, params string[] Aliases );
+namespace arnold.Utilities;
+
+public readonly record struct CommandRouter( string Name, string Description, Func<string[], int> Handler, IEnumerable<string> Aliases, IEnumerable<CommandRouter>? SubCommands = null );
 
 public class CommandProcessor( IEnumerable<CommandRouter> CommandRouters, string HelpText ) {
     /// <summary>
@@ -26,21 +29,51 @@ public class CommandProcessor( IEnumerable<CommandRouter> CommandRouters, string
         }
 
         foreach( var router in CommandRouters ) {
-            if( router.Aliases.Any( alias => alias.ToLower() == command ) ) {
-                return router.Handler( args.Skip(1).ToArray() );
+            if( router.Aliases.Any( alias => alias.Equals(command, StringComparison.CurrentCultureIgnoreCase)) ) {
+                return router.Handler([.. args.Skip(1)]);
             }
         }
 
-        return defaultPath(args.Skip(1).ToArray() );
+        return defaultPath([.. args.Skip(1)]);
     }
 
-    public static CommandRouter FallbackRouter = new CommandRouter(
+    public static string AssemblyName = System.Reflection.Assembly.GetAssembly( typeof(CommandProcessor) )!.GetName()!.Name!;
+    public static string GenerateHelpText( string[] command, string description, IEnumerable<CommandRouter>? routings = null, IEnumerable<ArgumentDefinition>? argumentMap = null ) {
+        routings = routings is null ? [] : routings.Where( routing => routing.Name != "fallback" );
+        argumentMap ??= [];
+
+        var helpBuilder = new StringBuilder($"{AssemblyName} {string.Join(" ", command)}");
+        helpBuilder.AppendLine(description);
+        
+        if( routings.Any() ) {
+            helpBuilder.AppendLine("  Commands");
+            helpBuilder.AppendLine("----------");
+            foreach( var routing in routings ) {
+                helpBuilder.AppendLine( $"  {routing.Name}: {routing.Description}" );
+                if( routing.Aliases.Any() ) helpBuilder.AppendLine("    " + string.Join(" ", routing.Aliases) );
+            }
+        }
+
+        if( argumentMap.Any() ) {
+            helpBuilder.AppendLine("  Arguments");
+            helpBuilder.AppendLine("-----------");
+            foreach(var argument in argumentMap ) {
+                helpBuilder.AppendLine($"  {argument.Name}: {argument.Description}" + (argument.Required ? " [required]" : "" ) );
+                if( argument.Aliases is not null && argument.Aliases.Any() ) {
+                    helpBuilder.AppendLine( "    " + string.Join(" ", argument.Aliases) );
+                }
+            }
+        }
+        return helpBuilder.ToString();
+    }
+
+    public static CommandRouter FallbackRouter = new(
         Name: "fallback",
         Description: "FAILS",
         Handler: args => {
             Console.WriteLine("Use the --help flag for available commands");
             return 0;
         },
-        Aliases: Array.Empty<string>()
+        Aliases: []
     );
 }
